make things a little more robust
