+++ integrations/oci/sdk/runtime/README.md
@@ -3,4 +3,4 @@
 This module offers integration for the OCI runtime environment. Most of the documentation for this be found under [helidon-integrations-oci-sdk](../).
 
 The [OciExtension](src/main/java/io/helidon/integrations/oci/sdk/runtime/OciExtension.java) class is the
-main access point for programmatically obtaining the global [OciConfig](target/generated-sources/annotations/io/helidon/integrations/oci/sdk/runtime/OciConfig.java) configuration. 
+main access point for programmatically obtaining the global [OciConfig](src/main/java/io/helidon/integrations/oci/sdk/runtime/OciConfigBlueprint.java) configuration. The javadoc also provides guidelines for configuration.
